requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
openpyxl==3.1.2
pydantic==2.6.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
lxml==5.1.0
sqlalchemy==2.0.25
requests==2.31.0
PyJWT==2.8.0
cryptography==42.0.5
    